Moved gui tests to tests package
